requirements.txt
apiview==1.3.25
asgiref==3.3.0
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.3
django-import-export==2.4.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
psycopg2==2.8.6
pylint==2.6.0
pylint-plugin-utils==0.6
pytz==2020.4
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.25.11
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
